requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
py-clob-client==0.19.0
